NPHW3 complete (works on Windows, not on Linux)

+++ HW3/modules/commands/logout.py
@@ -8,8 +8,13 @@ def logout(sv: SharedVariables):
     # check login status
     if sv.session_id == -1:
         raise RuntimeError(552072, "Please login first.")
-    # logout
+    # check chatroom status
+    chatroom_status = sv.sqlite_cursor.execute("SELECT status FROM chatrooms WHERE name = ? ", [sv.username]).fetchone()
+    if chatroom_status is not None and chatroom_status[0] == 'open':
+        raise RuntimeError(24526, "Please do “attach” and “leave-chatroom” first.")
+    # cleanup and logout
     sv.sqlite_cursor.execute("DELETE FROM sessions WHERE session_id = ? ", [sv.session_id])
     sv.tcp_conn.send(MSG_Encode(0))
     sv.session_id = -1
     sv.username = None
+    sv.chatroom_port = None
